azure_vault _to_ssm_param_store/requirements.txt
azure-identity==1.23.1
azure-keyvault-secrets==4.10.0
boto3==1.39.15